requirements.txt
crewai
langchain-google-genai
pandas
python-dotenv
streamlit
fpdf2
pypdf
pymupdf
python-docx
docx2txt
pytesseract
Pillow
psycopg2-binary
